+++ packages/components/src/Autocomplete/Autocomplete.css.d.ts
@@ -3,3 +3,5 @@ export const options: string;
 export const visible: string;
 export const option: string;
 export const active: string;
+export const icon: string;
+export const text: string;
