+++ packages/components/src/Autocomplete/Autocomplete.mdx
@@ -31,7 +31,7 @@ import { Autocomplete } from "@jobber/components/Autocomplete";
         { value: "6", label: "Enterprise-D" },
       ];
 
-    const [value, setValue] = useState("");
+    const [value, setValue] = useState();
 
     return (
       <Autocomplete
