+++ packages/components/src/Autocomplete/Option.ts
@@ -1,0 +1,6 @@
+type OptionValue = string | number;
+
+export interface Option {
+  value: OptionValue;
+  label: string;
+}
