+++ packages/components/src/Autocomplete/index.ts
@@ -1,1 +1,2 @@
-export * from "./Autocomplete";
+export { Autocomplete } from "./Autocomplete";
+export * from "./Option";
